% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{api_call}
\alias{api_call}
\title{Make general musiXmatch API call}
\usage{
api_call(method, body)
}
\arguments{
\item{method}{API function to call.
Options include artist.search, artist.albums.get, album.tracks.get.
More information can be found at https://developer.musixmatch.com/documentation/api-methods}
}
\description{
Make general musiXmatch API call
}

