% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_type}
\alias{check_type}
\title{Check type parameter is correct}
\usage{
check_type(type)
}
\arguments{
\item{type}{type of data to be returned}
}
\description{
Check type parameter is correct
}

