% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/album.R
\name{get_album}
\alias{get_album}
\title{Get information about the album}
\usage{
get_album(album_id, simplify = TRUE, ...)
}
\arguments{
\item{album_id}{ID of album on musiXmatch}
}
\description{
Get information about the album
}

