% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/album.R
\name{get_album_tracks}
\alias{get_album_tracks}
\title{Get a list of songs on the album}
\usage{
get_album_tracks(album_id, page_size = 100, simplify = TRUE, ...)
}
\arguments{
\item{album_id}{ID of album on musiXmatch}
}
\value{
a data.frame or list containing the data from the API call
}
\description{
Get a list of songs on the album
}

