% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{get_albums}
\alias{get_albums}
\title{Get album discography of an artist}
\usage{
get_albums(artist_id, page_size = 100, type = "data.frame", ...)
}
\arguments{
\item{artist_id}{MusiXmatcht artist id}

\item{number}{of pages in XML document}

\item{return}{type}
}
\value{
a data.frame or list containing the data from the API call
}
\description{
Get album discography of an artist
}

