% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/artist.R
\name{get_artist}
\alias{get_artist}
\title{Get information about an artist}
\usage{
get_artist(artist_id, simplify = TRUE, ...)
}
\arguments{
\item{artist_id}{musixmatch artist id}
}
\description{
Get information about an artist
}

