% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse_xml.R, R/simplify_xml.R
\name{get_artist_df}
\alias{get_artist_df}
\title{Returns XML parsed into easy to use data.frame}
\usage{
get_artist_df(xml)

get_artist_df(xml)
}
\arguments{
\item{xml}{an XML document with nodes artist_id, artist_name and artist_country}

\item{xml}{an XML document with nodes artist_id, artist_name and artist_country}
}
\description{
Returns XML parsed into easy to use data.frame

Returns XML parsed into easy to use data.frame
}

