% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/artist.R
\name{get_artist_related}
\alias{get_artist_related}
\title{Get other similar artists}
\usage{
get_artist_related(artist_id, page_size = 10, simplify = TRUE, ...)
}
\arguments{
\item{artist_id}{musixmatch artist_id}

\item{page_size}{}
}
\description{
Get other similar artists
}

