% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{get_lyrics}
\alias{get_lyrics}
\title{get lyrics from a song}
\usage{
get_lyrics(track_id, type = "data.frame", ...)
}
\arguments{
\item{track_id}{musiXmatch ID of song}
}
\value{
a data.frame or list containing the data from the API call
}
\description{
get lyrics from a song
}

