% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track.R
\name{get_track}
\alias{get_track}
\title{Get information about a track}
\usage{
get_track(track_id, simplify = TRUE, ...)
}
\arguments{
\item{track_id}{the musixmatch track ID}
}
\description{
Get information about a track
}

