% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track.R
\name{get_track_lyrics}
\alias{get_track_lyrics}
\title{get lyrics from a song}
\usage{
get_track_lyrics(track_id, simplify = TRUE, ...)
}
\arguments{
\item{track_id}{musiXmatch ID of song}
}
\value{
a data.frame or list containing the data from the API call
}
\description{
get lyrics from a song
}

