% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track.R
\name{get_track_snippet}
\alias{get_track_snippet}
\title{Get a lyrics snipped for a track.  A lyrics snippet is a very short representation of a song lyrics.
It’s usually twenty to a hundred characters long and it’s calculated extracting a sequence of words from the lyrics.}
\usage{
get_track_snippet(track_id, simplify = TRUE, ...)
}
\arguments{
\item{track_id}{the musixmatch track ID}
}
\description{
Get a lyrics snipped for a track.  A lyrics snippet is a very short representation of a song lyrics.
It’s usually twenty to a hundred characters long and it’s calculated extracting a sequence of words from the lyrics.
}

