% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track.R
\name{get_track_subtitle}
\alias{get_track_subtitle}
\title{Get subtitle of a track in LRC or DFXP format}
\usage{
get_track_subtitle(track_id, subtitle_format = "lrc", simplify = TRUE, ...)
}
\arguments{
\item{track_id}{musixmatch track ID}
}
\description{
Get subtitle of a track in LRC or DFXP format
}

