% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track.R
\name{post_track_lyrics}
\alias{post_track_lyrics}
\title{Submit a lyrics to Musixmatch database.}
\usage{
post_track_lyrics(track_id, lyrics_body, simplify = TRUE, ...)
}
\arguments{
\item{track_id}{the musixmatch track ID}

\item{lyrics_body}{the lyrics}
}
\description{
Submit a lyrics to Musixmatch database.
}

