% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track.R
\name{post_track_lyrics_feedback}
\alias{post_track_lyrics_feedback}
\title{Post feedback to a lyrics submission on musixmatch}
\usage{
post_track_lyrics_feedback(track_id, lyrics_id, feedback, simplify = TRUE,
  ...)
}
\arguments{
\item{track_id}{the musixmatch track_id}

\item{lyrics_id}{musixmatch lyrics id}

\item{feedback}{the feedback to be reported}
}
\description{
Post feedback to a lyrics submission on musixmatch
}

