% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track.R
\name{search_track}
\alias{search_track}
\title{Search for information about a track based on its artist, name, or lyrics}
\usage{
search_track(page_size = 100, simplify = TRUE, ...)
}
\arguments{
\item{q_track}{track name}

\item{q_artist}{artist name}

\item{q_lyrics}{song lyrics}
}
\description{
Search for information about a track based on its artist, name, or lyrics
}

