% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simplify_xml.R
\name{simplify_search_artist}
\alias{simplify_search_artist}
\title{Returns XML parsed into easy to use data.frame}
\usage{
simplify_search_artist(xml)
}
\arguments{
\item{xml}{an XML document with nodes artist_id, artist_name and artist_country}
}
\description{
Returns XML parsed into easy to use data.frame
}

