% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{get_artist}
\alias{get_artist}
\title{Search for artists in the database}
\usage{
get_artist(artist, type = "data.frame", ...)
}
\arguments{
\item{artist}{string of the artist to search for, e.g. "Slayer"}

\item{type}{type of object to return.}

\item{...}{other API parameters, e.g. artist_id, artist_mbid.
'data,frane' returns a data.frame of common elements of interest.
'list' will return the full XML results as a list}
}
\value{
a data.frame or list containing the data from the API call
}
\description{
Search for artists in the database
}
\examples{
get_artist('slayer')
get_artist('slayer',type='list')
}

