% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{get_full_list}
\alias{get_full_list}
\title{This returns a list containing all of the XML fields}
\usage{
get_full_list(xml)
}
\arguments{
\item{xml}{an XML document with nodes artist_id, artist_name and artist_country}
}
\value{
list of full XML document
}
\description{
This returns a list containing all of the XML fields
}

